refactor: deduplicate findExecutable into shared module function
Extract findExecutableInPath as internal module-level function in
WallpaperDataSource. Both YouTubeWallpaperDataSourceImpl and
WallpaperToolChecker now delegate to it. Also update PR description
to reflect always-check behavior.
